Move CleanUpButton to common
