+++ assets/js/common/CleanUpButton/CleanUpButton.jsx
