+++ assets/js/common/CleanUpButton/CleanUpButton.stories.jsx
