+++ assets/js/common/CleanUpButton/index.js
